requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
sse-starlette==2.1.2
httpx==0.28.1
python-dotenv==1.0.1
youtube-transcript-api==0.6.3
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain-huggingface==0.1.2
langchain-groq==0.2.2
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
